refactor(simulator): split CanvasApp into sibling canvas modules
Move the sidebar, diff view, iframe wrapper and tab header out of
CanvasApp.tsx into CanvasSidebar/DiffView/CanvasIframe/CanvasTabHeader,
and the payload helpers into canvasPayload.ts. CanvasApp keeps the
default export and all state; behaviour is unchanged.
